systemd: bump systemd version to 230
Since poky has been upgraded to the "morty" version, systemd version is
230 instead of 219. The bbappends recipe needs thus to be renamed
correctly. As for alsa-utils, the bbappend doesn't depend on a specific
systemd version as it only creates an network entry file to let an
interface be automatically configure via DHCP, so use a wildcard
instead.

As there is no issue with the udev mount propagation patch anymore in
this version, the bbappend that was in charge of removing this patch can
be dropped.