pjsip: clean configure step
Some content was appended to EXTRA_OECONF, but then the variable was
reassigned new content, so fix that and pass everything configure needs
to know in EXTRA_OECONF. Drop remaining exports as they are of no use.

Also, by default the configure step in Yocto is preceded by a `make
clean`, but pjsip breaks in that case. The previously chosen solution
was to override entirely the do_configure to 1) go the right directory,
2) call oe_runconf without cleaning first. Now that B is assigned the
right value, going to the right directory is handled. Yocto is also
aware that some packages don't support `make clean`, and allows us to
set CLEANBROKEN to skip this step.