qtmultimedia: make sure gstreamer plugins are shipped along it
First, it's not necessary to modify PACKAGECONFIG[gstreamer] to add imx
plugins at build-time dependency as this changes nothing to the way this
package is configure. Moreover, as this package don't link against these
plugins, there are not automatic runtime dependencies created.

What we really want to do is 1) enable gstreamer, 2) make sure gstreamer
plugins are shipped with qtmultimedia. To do so, we list these plugins
as RDEPENDS.