qtbase: enable fontconfig
This allows Qt to find system fonts through fontconfig, so there is no
need to put them in /usr/lib/fonts anymore (which is the path Qt expect
by default).

Also add a bbapend to revert changes introduced in meta-ts, at it
confuses fontconfig because fonts are not installed where they should.